Following best practices for separation of concerns, made the constructors and Initialize() methods take interfaces as parameter types.
Package implements both IVsPackage and IServiceProvider, so you can just cast from one to the other, it's akin to doing COM's QueryInterface() call.

+++ src/ExtensionManager.Shared/Commands/ImportCommand.cs
@@ -26,54 +26,130 @@ namespace ExtensionManager
 {
     internal sealed class ImportCommand
     {
-        private readonly Package _package;
-        private readonly ExtensionService _es;
+        /// <summary>
+        /// Reference to an instance of an object that implements the
+        /// <see cref="T:ExtensionManager.IExtensionService" /> interface.
+        /// </summary>
+        private readonly IExtensionService _extensionService;
 
-        private ImportCommand(Package package, OleMenuCommandService commandService, ExtensionService es)
+        /// <summary>
+        /// Reference to an instance of an object that implements the
+        /// <see cref="T:Microsoft.VisualStudio.Shell.Interop.IVsPackage" /> interface.
+        /// </summary>
+        private readonly IVsPackage _package;
+
+        private int _currentCount;
+
+        /// <summary>
+        /// Constructs a new instance of
+        /// <see cref="T:ExtensionManager.ImportCommand" /> and returns a reference
+        /// to it.
+        /// </summary>
+        /// <param name="package">
+        /// (Required.) Reference to an instance of an object that implements the
+        /// <see cref="T:Microsoft.VisualStudio.Shell.Interop.IVsPackage" /> interface.
+        /// </param>
+        /// <param name="commandService">
+        /// (Required.) Reference to an instance of an object that implements the
+        /// <see cref="T:System.ComponentModel.Design.IMenuCommandService" /> interface.
+        /// </param>
+        /// <param name="extensionService">
+        /// (Required.) Reference to an instance of an object that implements the
+        /// <see cref="T:ExtensionManager.IExtensionService" /> interface.
+        /// </param>
+        /// null
+        /// <exception cref="T:System.ArgumentNullException">
+        /// Thrown if the any of the
+        /// required parameters, <paramref name="package" />,
+        /// <paramref name="commandService" />, or <paramref name="extensionService" />,
+        /// are passed a <see langword="null" /> value.
+        /// </exception>
+        private ImportCommand(IVsPackage package,
+            IMenuCommandService commandService,
+            IExtensionService extensionService)
         {
-            _package = package ?? throw new ArgumentNullException(nameof(package));
-            _es = es;
+            if (commandService == null)
+                throw new ArgumentNullException(nameof(commandService));
 
-            var cmdId = new CommandID(PackageGuids.guidExportPackageCmdSet, PackageIds.ImportCmd);
+            _package = package ??
+                       throw new ArgumentNullException(nameof(package));
+            _extensionService = extensionService ??
+                                throw new ArgumentNullException(
+                                    nameof(extensionService)
+                                );
+
+            var cmdId = new CommandID(
+                PackageGuids.guidExportPackageCmdSet, PackageIds.ImportCmd
+            );
             var cmd = new MenuCommand(Execute, cmdId);
             commandService.AddCommand(cmd);
         }
 
         public static ImportCommand Instance { get; private set; }
 
-        private IServiceProvider ServiceProvider => _package;
+        private IServiceProvider ServiceProvider
+            => (IServiceProvider)_package;
 
-        private int _currentCount;
         private int EntriesCount { get; set; }
 
-        public static void Initialize(AsyncPackage package, OleMenuCommandService commandService, IExtensionService es)
+        /// <summary>
+        /// Creates and initializes a new instance of
+        /// <see cref="T:ExtensionManager.ImportCommand" /> and sets the
+        /// <see cref="P:ExtensionManager.ImportCommand.Instance" /> property to a
+        /// reference to it.
+        /// </summary>
+        /// <param name="package">
+        /// (Required.) Reference to an instance of an object that implements the
+        /// <see cref="T:Microsoft.VisualStudio.Shell.Interop.IVsPackage" /> interface.
+        /// </param>
+        /// <param name="commandService">
+        /// (Required.) Reference to an instance of an object that implements the
+        /// <see cref="T:System.ComponentModel.Design.IMenuCommandService" /> interface.
+        /// </param>
+        /// <param name="extensionService">
+        /// (Required.) Reference to an instance of an object that implements the
+        /// <see cref="T:ExtensionManager.IExtensionService" /> interface.
+        /// </param>
+        /// null
+        /// <exception cref="T:System.ArgumentNullException">
+        /// Thrown if the any of the
+        /// required parameters, <paramref name="package" />,
+        /// <paramref name="commandService" />, or <paramref name="extensionService" />,
+        /// are passed a <see langword="null" /> value.
+        /// </exception>
+        public static void Initialize(IVsPackage package,
+            IMenuCommandService commandService, IExtensionService extensionService)
         {
-            Instance = new ImportCommand(package, commandService, es);
+            Instance = new ImportCommand(package, commandService, extensionService);
         }
 
         private void Execute(object sender, EventArgs e)
         {
             ThreadHelper.ThrowIfNotOnUIThread();
 
-            if (!TryGetFilePath(out var filePath))
-            {
-                return;
-            }
+            if (!TryGetFilePath(out var filePath)) return;
 
             var manifest = Manifest.FromFile(filePath);
-            manifest.MarkSelected(_es.GetInstalledExtensions());
+            manifest.MarkSelected(_extensionService.GetInstalledExtensions());
 
             var dialog = ImportWindow.Open(manifest.Extensions, Purpose.Import);
 
             if (dialog.DialogResult == true && dialog.SelectedExtension.Any())
             {
                 var installSystemWide = dialog.InstallSystemWide;
-                var toInstall = dialog.SelectedExtension.Select(ext => ext.ID).ToList();
+                var toInstall = dialog.SelectedExtension.Select(ext => ext.ID)
+                                      .ToList();
 
-                var repository = ServiceProvider.GetService(typeof(SVsExtensionRepository)) as IVsExtensionRepository;
+                var repository =
+                    ServiceProvider.GetService(typeof(SVsExtensionRepository))
+                        as IVsExtensionRepository;
                 Assumes.Present(repository);
 
-                IEnumerable<GalleryEntry> marketplaceEntries = repository.GetVSGalleryExtensions<GalleryEntry>(toInstall, 1033, false).Where(x => x.DownloadUrl != null);
+                var marketplaceEntries = repository
+                                         .GetVSGalleryExtensions<GalleryEntry>(
+                                             toInstall, 1033, false
+                                         )
+                                         .Where(x => x.DownloadUrl != null);
                 var tempDir = PrepareTempDir();
 
                 var dte = ServiceProvider.GetService(typeof(DTE)) as DTE;
@@ -82,11 +158,18 @@ namespace ExtensionManager
                 dte.StatusBar.Text = "Downloading extensions...";
 
                 HasRootSuffix(out var rootSuffix);
-                ThreadHelper.JoinableTaskFactory.Run(() => DownloadExtensionsAsync(installSystemWide, marketplaceEntries, tempDir, dte, rootSuffix));
+                ThreadHelper.JoinableTaskFactory.Run(
+                    () => DownloadExtensionsAsync(
+                        installSystemWide, marketplaceEntries, tempDir, dte,
+                        rootSuffix
+                    )
+                );
             }
         }
 
-        private async Task DownloadExtensionsAsync(bool installSystemWide, IEnumerable<GalleryEntry> marketplaceEntries, string tempDir, DTE dte, string rootSuffix)
+        private async Task DownloadExtensionsAsync(bool installSystemWide,
+            IEnumerable<GalleryEntry> marketplaceEntries, string tempDir,
+            DTE dte, string rootSuffix)
         {
             ServicePointManager.DefaultConnectionLimit = 100;
             EntriesCount = marketplaceEntries.Count();
@@ -94,7 +177,8 @@ namespace ExtensionManager
             await DownloadExtensionAsync(marketplaceEntries, tempDir, dte);
 
             await ThreadHelper.JoinableTaskFactory.SwitchToMainThreadAsync();
-            dte.StatusBar.Text = "Extensions downloaded. Starting VSIX Installer...";
+            dte.StatusBar.Text =
+                "Extensions downloaded. Starting VSIX Installer...";
             InvokeVsixInstaller(tempDir, rootSuffix, installSystemWide);
         }
 
@@ -108,7 +192,7 @@ namespace ExtensionManager
                 sfd.FileName = "extensions";
                 sfd.Filter = "VSEXT File|*.vsext";
 
-                DialogResult result = sfd.ShowDialog();
+                var result = sfd.ShowDialog();
 
                 if (result == DialogResult.OK)
                 {
@@ -122,54 +206,67 @@ namespace ExtensionManager
 
         private static string PrepareTempDir()
         {
-            var tempDir = Path.Combine(Path.GetTempPath(), nameof(ExtensionManager));
+            var tempDir = Path.Combine(
+                Path.GetTempPath(), nameof(ExtensionManager)
+            );
 
-            if (Directory.Exists(tempDir))
-            {
-                Directory.Delete(tempDir, true);
-            }
+            if (Directory.Exists(tempDir)) Directory.Delete(tempDir, true);
 
             Directory.CreateDirectory(tempDir);
             return tempDir;
         }
 
-        private void InvokeVsixInstaller(string tempDir, string rootSuffix, bool installSystemWide)
+        private void InvokeVsixInstaller(string tempDir, string rootSuffix,
+            bool installSystemWide)
         {
             var process = Process.GetCurrentProcess();
             var dir = Path.GetDirectoryName(process.MainModule.FileName);
             var exe = Path.Combine(dir, "VSIXInstaller.exe");
             var configuration = new SetupConfiguration() as ISetupConfiguration;
             var adminSwitch = installSystemWide ? "/admin" : string.Empty;
-            ISetupInstance instance = configuration.GetInstanceForCurrentProcess();
-            IEnumerable<string> vsixFiles = Directory.EnumerateFiles(tempDir, "*.vsix").Select(f => Path.GetFileName(f));
+            var instance = configuration.GetInstanceForCurrentProcess();
+            var vsixFiles = Directory.EnumerateFiles(tempDir, "*.vsix")
+                                     .Select(f => Path.GetFileName(f));
 
             var start = new ProcessStartInfo {
                 FileName = exe,
-                Arguments = $"{string.Join(" ", vsixFiles)} /instanceIds:{instance.GetInstanceId()} {adminSwitch}",
+                Arguments =
+                    $"{string.Join(" ", vsixFiles)} /instanceIds:{instance.GetInstanceId()} {adminSwitch}",
                 WorkingDirectory = tempDir,
-                UseShellExecute = false,
+                UseShellExecute = false
             };
 
             if (!string.IsNullOrEmpty(rootSuffix))
-            {
                 start.Arguments += $" /rootSuffix:{rootSuffix}";
-            }
 
             Process.Start(start);
         }
 
-        private Task DownloadExtensionAsync(IEnumerable<GalleryEntry> entries, string dir, DTE dte)
+        private Task DownloadExtensionAsync(IEnumerable<GalleryEntry> entries,
+            string dir, DTE dte)
         {
-            return Task.WhenAll(entries.Select(entry => Task.Run(() => DownloadExtensionFileAsync(entry, dir, dte))).ToArray());
+            return Task.WhenAll(
+                entries.Select(
+                           entry => Task.Run(
+                               () => DownloadExtensionFileAsync(entry, dir, dte)
+                           )
+                       )
+                       .ToArray()
+            );
         }
 
-        private async Task DownloadExtensionFileAsync(GalleryEntry entry, string dir, DTE dte)
+        private async Task DownloadExtensionFileAsync(GalleryEntry entry,
+            string dir, DTE dte)
         {
-            var localPath = Path.Combine(dir, CreateMD5(entry.DownloadUrl) + ".vsix");
+            var localPath = Path.Combine(
+                dir, CreateMD5(entry.DownloadUrl) + ".vsix"
+            );
 
             using (var client = new WebClient())
             {
-                await client.DownloadFileTaskAsync(entry.DownloadUrl, localPath);
+                await client.DownloadFileTaskAsync(
+                    entry.DownloadUrl, localPath
+                );
             }
 
             await UpdateProgressAsync(dte);
@@ -178,7 +275,8 @@ namespace ExtensionManager
         private async Task UpdateProgressAsync(DTE dte)
         {
             await ThreadHelper.JoinableTaskFactory.SwitchToMainThreadAsync();
-            dte.StatusBar.Text = $"Downloaded {Interlocked.Increment(ref _currentCount)} of {EntriesCount} extensions...";
+            dte.StatusBar.Text =
+                $"Downloaded {Interlocked.Increment(ref _currentCount)} of {EntriesCount} extensions...";
             await TaskScheduler.Default;
         }
 
@@ -186,32 +284,35 @@ namespace ExtensionManager
         {
             try
             {
-                using (MD5 md5 = MD5.Create())
+                using (var md5 = MD5.Create())
                 {
-                    byte[] inputBytes = Encoding.ASCII.GetBytes(input);
-                    byte[] hashBytes = md5.ComputeHash(inputBytes);
-                    return BitConverter.ToString(hashBytes).Replace("-", string.Empty).ToLower();
+                    var inputBytes = Encoding.ASCII.GetBytes(input);
+                    var hashBytes = md5.ComputeHash(inputBytes);
+                    return BitConverter.ToString(hashBytes)
+                                       .Replace("-", string.Empty)
+                                       .ToLower();
                 }
             }
-            catch (Exception)
-            {
-            }
+            catch (Exception) { }
 
             return null;
         }
+
         public static bool HasRootSuffix(out string rootSuffix)
         {
             ThreadHelper.ThrowIfNotOnUIThread();
 
             rootSuffix = null;
 
-            if (Microsoft.VisualStudio.Shell.ServiceProvider.GlobalProvider.GetService(typeof(SVsAppCommandLine)) is IVsAppCommandLine appCommandLine)
-            {
-                if (ErrorHandler.Succeeded(appCommandLine.GetOption("rootsuffix", out var hasRootSuffix, out rootSuffix)))
-                {
+            if (Microsoft.VisualStudio.Shell.ServiceProvider.GlobalProvider
+                         .GetService(typeof(SVsAppCommandLine)) is
+                IVsAppCommandLine appCommandLine)
+                if (ErrorHandler.Succeeded(
+                        appCommandLine.GetOption(
+                            "rootsuffix", out var hasRootSuffix, out rootSuffix
+                        )
+                    ))
                     return hasRootSuffix != 0;
-                }
-            }
 
             return false;
         }
